Struct protocoll::map::VecSortedMap

pub struct VecSortedMap<K, V>(_);

an array-map sorted by key.

does not support entry; see Map::update for the same functionality. MapMut functions are not much more efficient than Map for this data structure.

Methods

impl<K, V> VecSortedMap<K, V> where K: Ord

fn new() -> Self

fn with_capacity(c: usize) -> Self

fn capacity(&self) -> usize

fn reserve(&mut self, c: usize)

fn shrink_to_fit(&mut self)

fn clear(&mut self)

fn contains_key<Q: ?Sized>(&self, k: &Q) -> bool where K: Borrow<Q>, Q: Ord

O(log(len))

fn get<Q: ?Sized>(&self, k: &Q) -> Option<&V> where K: Borrow<Q>, Q: Ord

O(log(len))

fn get_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut V> where K: Borrow<Q>, Q: Ord

O(log(len))

fn insert(&mut self, k: K, v: V) -> Option<V>

O(log(len)) when k already exists. O(len) for inserting a new entry, caused by shifting all entries after it, which can be avoided by always inserting in order.

fn remove<Q: ?Sized>(&mut self, k: &Q) -> Option<V> where K: Borrow<Q>, Q: Ord

O(log(len)) when k does not exist. O(len) for removing an entry, because of the need for shifting all entries after it.

fn append(&mut self, other: &mut VecSortedMap<K, V>)

impl<K, V> VecSortedMap<K, V>

fn view_content<'a>(&'a self) -> &'a Vec<(K, V)>

a view for the underlying vec. &self methods for Vec such as get and split can be accessed through this.

fn iter(&self) -> Iter<(K, V)>

iterate over the underlying vec. note: iterator element type is not (&K,&V) but rather &(K,V). iter_mut is not supported for this collection. see update_all_mut for the same functionality.

fn len(&self) -> usize

fn is_empty(&self) -> bool