Struct protocoll::set::VecSortedSet

pub struct VecSortedSet<T>(_);

an array-set.

Methods

impl<T> VecSortedSet<T> where T: Ord

fn new() -> Self

fn with_capacity(c: usize) -> Self

fn capacity(&self) -> usize

fn reserve(&mut self, c: usize)

fn shrink_to_fit(&mut self)

fn clear(&mut self)

fn contains<Q: ?Sized>(&self, e: &Q) -> bool where T: Borrow<Q>, Q: Ord

O(log(len))

fn get<Q: ?Sized>(&self, e: &Q) -> Option<&T> where T: Borrow<Q>, Q: Ord

O(log(len))

fn get_mut<Q: ?Sized>(&mut self, e: &Q) -> Option<&mut T> where T: Borrow<Q>, Q: Ord

O(log(len))

fn insert(&mut self, e: T) -> Option<T>

O(log(len)) when e already exists. O(len) for inserting a new element, caused by shifting all elements after it, which can be avoided by always inserting in order.

fn remove<Q: ?Sized>(&mut self, e: &Q) -> Option<T> where T: Borrow<Q>, Q: Ord

O(log(len)) when e does not exist. O(len) for removing an element, because of the need for shifting all elements after it.

fn append(&mut self, other: &mut VecSortedSet<T>)

impl<T> VecSortedSet<T>

fn view_content<'a>(&'a self) -> &'a Vec<T>

a view for the underlying vec. &self methods for Vec such as get and split can be accessed through this.

fn iter(&self) -> Iter<T>

iterate over the underlying vec.

fn len(&self) -> usize

fn is_empty(&self) -> bool